+++ node/examples/account.ts
@@ -2,13 +2,6 @@ import 'dotenv/config'
 
 import { SttClient } from '@aristech-org/stt-client'
 
-const auth = process.env.TOKEN && process.env.SECRET ? { token: process.env.TOKEN, secret: process.env.SECRET } : undefined
-
-const client = new SttClient({
-  host: process.env.HOST,
-  ssl: Boolean(auth) || Boolean(process.env.ROOT_CERT) || process.env.SSL === 'true',
-  rootCert: process.env.ROOT_CERT,
-  auth,
-})
+const client = new SttClient()
 const info = await client.accountInfo()
 console.log(info)