+++ node/examples/nlpProcess.ts
@@ -2,14 +2,7 @@ import 'dotenv/config'
 
 import { SttClient } from '@aristech-org/stt-client'
 
-const auth = process.env.TOKEN && process.env.SECRET ? { token: process.env.TOKEN, secret: process.env.SECRET } : undefined
-
-const client = new SttClient({
-  host: process.env.HOST,
-  ssl: Boolean(auth) || Boolean(process.env.ROOT_CERT) || process.env.SSL === 'true',
-  rootCert: process.env.ROOT_CERT,
-  auth,
-})
+const client = new SttClient()
 
 // To see which server configs and functions within those servers are available, run the nlpFunctions.ts example
 const serverConfig = process.env.NLP_SERVER_CONFIG || 'default'