+++ node/examples/streaming.ts
@@ -18,13 +18,7 @@ if (!fileName) {
 }
 
 // Create a client and a recognition stream
-const auth = process.env.TOKEN && process.env.SECRET ? { token: process.env.TOKEN, secret: process.env.SECRET } : undefined
-const client = new SttClient({
-  host: process.env.HOST,
-  ssl: Boolean(auth) || Boolean(process.env.ROOT_CERT) || process.env.SSL === 'true',
-  rootCert: process.env.ROOT_CERT,
-  auth,
-})
+const client = new SttClient()
 const stream = client.recognize({
   specification: {
     // When streaming, we need to provide the sample rate, because the data are expected to be raw PCM data without a header