+++ python/examples/nlpFunctions.py
@@ -1,6 +1,7 @@
 from aristech_stt_client import SttClient
-from utils import host, auth_token, auth_secret, root_cert, ssl
+from dotenv import load_dotenv
 
-client = SttClient(host=host, ssl=ssl, root_cert=root_cert, auth_token=auth_token, auth_secret=auth_secret)
+load_dotenv()
+client = SttClient()
 response = client.list_nlp_functions()
 print(response)